Target field activities per module and scheme

The data frame mod_scheme_targetfieldactivity contains the main field activities that are related to the target variable(s) of a module and scheme: the so-called 'target field activities'.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

mod_scheme_targetfieldactivity is indirectly derived from mod_scheme_variableset_mainfieldactivity.

Restriction by domains and sampling frames

This object has been further restricted by module_domain_schemes. This means that the current spatial distribution of target populations of the module's schemes within the module's domain(s) is taken into account, using the scheme's spatial sampling frame. When the domains of a module represent a relatively small area, it is possible:

for target populations: that some types (that were initially targeted by the module) are not present in that area and consequently dropped from the module's target populations

for schemes: that some schemes targeted by the module are not applicable, i.e. when no types of its target population are present in that area

Also, note that for MHQ schemes, module_domain_schemes only considers – on a domain basis – the schemes with sampling approaches, not the census approaches, and this also leads to a further restriction.

See also