Spatial sample sizes of strata per module x domain x scheme x panel set

The data frame module_domain_scheme_ps_stratum_sample_size contains the spatial sample size per stratum in each module x domain x scheme x panel set.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

module_domain_scheme_ps_stratum_sample_size plays a key role in the spatial sampling procedure. It is derived from submodule_domain_scheme_ps_designattr, it no longer distinguishes between submodules (see submodules) and takes into account finite population correction.

Note

This object still has the full target population for each module x domain x scheme in the case of non-core schemes, but the sample sizes will generally be missing values since they are not designed: only the locations that are common between the sample of a related core scheme and the sampling frame of the non-core scheme can be attributed to the non-core scheme (insofar the core scheme schedules data collection that makes sense to the non-core scheme target variables). So this sample can be a subset of the sample of the core scheme. Since it causes no extra fieldwork, we don't explicitly consider these sample sizes and samples. (There's still an exception possible: we exceptionally attribute an extra sample size in non-core MNE schemes in order to adopt an extra type (which must belong to the non-core scheme's target population) that is not covered by a corresponding core MNE scheme, if it makes an important difference to that non-core scheme.)

See also