Modules

The data frame modules defines the modules, their names, shortnames and codes, as well as several temporal attributes.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

A module is like a project, that aims at realizing some well-defined part of MNE + MHQ. It determines which schemes, which domains and which part of the target populations are in scope for the module. For the schemes and target populations this happens indirectly via submodules; for the domains this is via module_domains.

Several objects with scheme properties have a module column to make these properties module-dependent.

See also