Designed fieldwork calendar per scheme x module combo x panel set x spatial poststratum

The data frame scheme_moco_ps_spsubset_fag_stratum_sppost_spsamples_calendar represents the designed sampling unit calendar of the distinguished field activity groups (FAGs), after cleaning within, but not between, levels of scheme x module combo x panel set x spatial poststratum.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

scheme_moco_ps_spsubset_fag_stratum_sppost_spsamples_calendar combines the panel calendar and the panel memberships generated in the REP workflow, which have been (indirectly) based on mod_scheme_ps_fag_paneldesign and scheme_moco_ps_stratum_sppost_spsamples. Panels and FAG occasions have been dropped if the scheduled calendar starts at a later point in time. Also some superfluous occasions of specific auxiliary FAGs have been dropped.

See also