Target panel designs per scheme x panel set x module combo

The data frame scheme_ps_target_paneldesign contains only the panel designs (from mod_scheme_ps_fag_paneldesign) for the so called 'target field activity groups' (target FAGs), i.e. the FAGs containing the target field activity (see mod_scheme_targetfieldactivity). These panel designs can be called target panel designs and the corresponding panels will be called 'targetpanels'. Importantly, the module level is gone: instead, modules with the same target panel design are combined as 'module combos' at the scheme x panel set level.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

scheme_ps_target_paneldesign is derived from mod_scheme_targetfieldactivity, mod_scheme_ps_fag_paneldesign and mod_scheme_field_activity (to link field activity and field activity groups).

The target variables limit the main field activities of a module x scheme to a great extent, since the latter also comprise variables for explanation & quality control. (Remember: the term 'main field activity' is to be understood from the viewpoint of a variable (set) in a module x scheme, not just the module x scheme; see mod_scheme_variableset_mainfieldactivity.)

For each scheme x panel set, the modules (in mod_scheme_ps_fag_paneldesign) have been collapsed to module combos (module combinations) if the corresponding target panel design is the same.

So, beware that the module combo can only be interpreted as a unique panel design in combination with scheme x panel set.

scheme_ps_target_paneldesign is the object in the REP-workflow where the module combos are effectively generated and defined.

However, see the derived data frame mod_scheme_yrs_moco_ps for a convenient object that links modules and module combos within scheme x panel set.

The concept of module combos is important to the spatial sampling procedure: spatial sampling for a scheme x panel set is to be combined over active modules when their target panel design and domain is the same. Also with overlapping domains, the sampling procedure must take into account the relation between the samples from both domains.

Note that the 'spatial subset' from mod_scheme_ps_fag_paneldesign is gone. That is because for the main field activities the spatial subsets don't have differences in the corresponding panel designs.

See also