Panel designs of measurements (data occurrence) per scheme x panel set

The data frame scheme_ps_variableset_paneldesign contains the panel design of measurements at the level of a variable set in a scheme x panel set. With 'measurements' we mean the data occurrence, not the human activity. The associated 'main field activity' is included though to make clear how the data are collected. All variables are covered, so this is different from the scope of scheme_ps_target_paneldesign, where only the 'target field activities' (main field activities for target variables) are considered.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

scheme_ps_variableset_paneldesign has statistical relevance and it is (should be) in line with relevant parts of mod_scheme_ps_fag_paneldesign, but it is not used further in the REP workflow.

scheme_ps_variableset_paneldesign uses mod_scheme_variableset_mainfieldactivity as its base.

Notable differences with mod_scheme_ps_fag_paneldesign are that:

automated frequent measurements are reflected as 'frequent revisit' while fieldwork is much less frequent;

currently no distinction between modules is made;

only core schemes are considered;

no installation or maintenance is involved, since those are not main field activities (for a variable set).

See also