Named sets of schemes with target populations

The data frame scheme_targetpop_sets defines different collections of schemes with associated target populations. Each submodule in the submodules object refers to a single 'scheme + target population set', i.e. to one row of this data frame.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

The target populations are given at type level and at stratum level. The stratum level target population follows from the type level, as all strata of each involved type are included.

These target populations cannot exceed the target populations defined by n2khabmon::read_scheme_types(), but they can be subsets considered by a specific (sub)module, which is their true aim.

Note that scheme_targetpop_sets can list more sets than actually used by the active modules and submodules.

See also