Spatial sampling frames

The data frame sp_samplingframes defines spatial sampling frames that can be shared among different schemes and modules, but which are generally still 'too large' for a specific scheme in a specific module. The object scheme_sampling_frame defines which spatial sampling frame should be further processed for use in a specific scheme.

Format

Data frame.

Source

The REP (revisitplan) workflow. See https://github.com/inbo/n2khab-mne-designs/tree/revisitplan/100_design_common/010_revisitplan.

Details

These spatial sampling frames are the result of applying some special spatial restrictions to the base sampling frame as implied by n2khabmon::read_scheme_types(). To obtain the ultimate spatial sampling frame for a specific scheme in a specific module, these sampling frames must still be subsetted using the target population and domain implied by a scheme in a module. A summarized result of that final step can be found in submod_dom_scheme_ssf_stratum_nunits.

See also